edit main files to be more readable and clean
